refactor board and scoreboard components
